students/k3343/kursoviks/Andreeva_Ekaterina/hotel-backend/requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
distlib==0.3.0
Django==3.0
django-allauth==0.42.0
django-ckeditor==5.9.0
django-cors-headers==2.2.0
django-filter==2.2.0
django-filters==0.2.1
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-rest-swagger==2.1.2
django-summernote==0.8.8.6
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jsonapi==2.4.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==1.1.5
filelock==3.0.12
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.1.2
pipenv==2018.11.26
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.3.0
simplejson==3.16.0
six==1.11.0
sqlparse==0.3.1
uritemplate==3.0.0
urllib3==1.23
virtualenv==20.0.21
virtualenv-clone==0.5.4
